Addressed scaling issue for physical etch-a-sketch.
Basically, force the virtual bounding box to be the same dimensions
as the physical bounding box. The assumption is that the
physical etch-a-sketch will then compute the same layout given
the user's offset_left, offset_bottom, and zoom values as that on
the HTML canvas.

Next up: Add the actual Etch command.